feat: improve default layout
