+++ lib/tiny_admin/views/actions/index.rb
@@ -5,7 +5,7 @@ module TinyAdmin
     module Actions
       class Index < DefaultLayout
         attr_reader :current_page, :fields, :pages, :prepare_record, :records
-        attr_accessor :actions, :filters
+        attr_accessor :actions, :filters, :query_string
 
         def setup_pagination(current_page:, pages:)
           @current_page = current_page
