+++ lib/tiny_admin/views/basic_layout.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module TinyAdmin
+  module Views
+    class BasicLayout < Phlex::HTML
+      include Utils
+
+      def components
+        settings.components
+      end
+
+      def update_attributes(attributes)
+        attributes.each do |key, value|
+          send("#{key}=", value)
+        end
+      end
+    end
+  end
+end
