+++ lib/tiny_admin/views/default_layout.rb
@@ -2,30 +2,19 @@
 
 module TinyAdmin
   module Views
-    class DefaultLayout < Phlex::HTML
-      include Utils
-
-      attr_accessor :context, :messages, :options, :query_string, :title
-
-      def update_attributes(attributes)
-        attributes.each do |key, value|
-          send("#{key}=", value)
-        end
-      end
+    class DefaultLayout < BasicLayout
+      attr_accessor :context, :messages, :options, :title
 
       def template(&block)
-        @messages ||= {}
-        items = options&.include?(:no_menu) ? [] : settings.navbar
-
         doctype
         html {
           render components[:head].new(title, style_links: style_links, extra_styles: settings.extra_styles)
 
           body(class: body_class) {
-            render components[:navbar].new(current_slug: context&.slug, root: settings.root, items: items)
+            render components[:navbar].new(current_slug: context&.slug, root: settings.root, items: navbar_items)
 
             main_content {
-              render components[:flash].new(messages: messages)
+              render components[:flash].new(messages: messages || {})
               yield_content(&block)
             }
 
@@ -36,24 +25,12 @@ module TinyAdmin
 
       private
 
-      def components
-        settings.components
-      end
-
-      def style_links
-        settings.style_links || [
-          # Bootstrap CDN
-          {
-            href: 'https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css',
-            rel: 'stylesheet',
-            integrity: 'sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65',
-            crossorigin: 'anonymous'
-          }
-        ]
-      end
-
       def body_class
         "module-#{self.class.to_s.split('::').last.downcase}"
+      end
+
+      def navbar_items
+        options&.include?(:no_menu) ? [] : settings.navbar
       end
 
       def main_content
@@ -70,10 +47,20 @@ module TinyAdmin
         end
       end
 
-      def render_scripts
-        return unless settings.scripts
+      def style_links
+        settings.style_links || [
+          # Bootstrap CDN
+          {
+            href: 'https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css',
+            rel: 'stylesheet',
+            integrity: 'sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65',
+            crossorigin: 'anonymous'
+          }
+        ]
+      end
 
-        settings.scripts.each do |script_attrs|
+      def render_scripts
+        (settings.scripts || []).each do |script_attrs|
           script(**script_attrs)
         end
       end
